+++ resources/views/bootstrap-components/media/audio.blade.php
@@ -5,6 +5,6 @@
         {{ classTag('audio-component', $componentClass) }}
         {{ htmlAttributes($componentHtmlAttributes) }}>
         <source {{ htmlAttributes($src ? 'src="'.$src.'"' : null) }}>
-        {{ trans('bootstrap-components::bootstrap-components.notification.audio') }}
+        @lang('bootstrap-components::bootstrap-components.notification.audio')
     </audio>
 </div>