+++ resources/views/bootstrap-components/media/video.blade.php
@@ -6,6 +6,6 @@
         {{ htmlAttributes($componentHtmlAttributes) }}
         {{ htmlAttributes($poster ? 'poster="'.$poster.'"' : null) }}>
         <source {{ htmlAttributes($src ? 'src="'.$src.'"' : null) }}>
-        {{ trans('bootstrap-components::bootstrap-components.notification.video') }}
+        @lang('bootstrap-components::bootstrap-components.notification.video')
     </video>
 </div>