+++ tests/Unit/Form/DatetimeTest.php
@@ -210,11 +210,8 @@ class DatetimeTest extends BootstrapComponentsTestCase
         $configLegend = 'test-config-legend';
         config()->set('bootstrap-components.form.datetime.legend', $configLegend);
         $html = bsDatetime()->name('name')->toHtml();
-        $this->assertStringContainsString(
-            '<small id="datetime-local-name-legend" class="form-text text-muted">bootstrap-components::'
-            . $configLegend . '</small>',
-            $html
-        );
+        $this->assertStringContainsString('datetime-local-name-legend', $html);
+        $this->assertStringContainsString($configLegend, $html);
     }
 
     public function testSetLegend()
@@ -223,37 +220,24 @@ class DatetimeTest extends BootstrapComponentsTestCase
         $customLegend = 'test-custom-legend';
         config()->set('bootstrap-components.form.datetime.legend', $configLegend);
         $html = bsDatetime()->name('name')->legend($customLegend)->toHtml();
-        $this->assertStringContainsString(
-            '<small id="datetime-local-name-legend" class="form-text text-muted">'
-            . $customLegend . '</small>',
-            $html
-        );
-        $this->assertStringNotContainsString(
-            '<small id="datetime-local-name-legend" class="form-text text-muted">bootstrap-components::'
-            . $configLegend . '</small>',
-            $html
-        );
+        $this->assertStringContainsString('datetime-local-name-legend', $html);
+        $this->assertStringContainsString($customLegend, $html);
+        $this->assertStringNotContainsString($configLegend, $html);
     }
 
     public function testNoLegend()
     {
         config()->set('bootstrap-components.form.datetime.legend', null);
         $html = bsDatetime()->name('name')->toHtml();
-        $this->assertStringNotContainsString(
-            '<small id="datetime-local-name-legend" class="form-text text-muted">"',
-            $html
-        );
+        $this->assertStringNotContainsString('datetime-local-name-legend', $html);
     }
 
     public function testHideLegend()
     {
         $configLegend = 'test-config-legend';
         config()->set('bootstrap-components.form.datetime.legend', $configLegend);
-        $html = bsDatetime()->name('name')->hideLegend()->toHtml();
-        $this->assertStringNotContainsString(
-            '<small id="datetime-local-name-legend" class="form-text text-muted">',
-            $html
-        );
+        $html = bsDatetime()->name('name')->legend(false)->toHtml();
+        $this->assertStringNotContainsString('datetime-local-name-legend', $html);
     }
 
     public function testSetWrongValue()
@@ -314,7 +298,7 @@ class DatetimeTest extends BootstrapComponentsTestCase
 
     public function testHideLabel()
     {
-        $html = bsDatetime()->name('name')->hideLabel()->toHtml();
+        $html = bsDatetime()->name('name')->label(false)->toHtml();
         $this->assertStringNotContainsString(
             '<label for="date-name">validation.attributes.name</label>',
             $html
@@ -346,13 +330,19 @@ class DatetimeTest extends BootstrapComponentsTestCase
         $this->assertStringContainsString('placeholder="validation.attributes.name"', $html);
     }
 
+    public function testHidePlaceholder()
+    {
+        $html = bsDatetime()->name('name')->placeholder(false)->toHtml();
+        $this->assertStringNotContainsString('placeholder="', $html);
+    }
+
     public function testSuccess()
     {
         $messageBag = app(MessageBag::class)->add('other_name', null);
         $html = bsDatetime()->name('name')->render(['errors' => $messageBag]);
         $this->assertStringContainsString('<div class="valid-feedback d-block">', $html);
         $this->assertStringContainsString(
-            trans('bootstrap-components::bootstrap-components.notification.validation.success'),
+            __('bootstrap-components::bootstrap-components.notification.validation.success'),
             $html
         );
     }